Move auth-server.js to aspsp-authorisation-server/token.js
Create a directory to store files for mocking aspsp-authorisation-server.
