+++ lib/app.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { OBAccountPaymentServiceProviders } = require('./ob-directory');
-const { authServer } = require('./auth-server');
+const { authServer } = require('./aspsp-authorisation-server/token');
 const { accountRequests } = require('./account-requests');
 const { openIdConfig } = require('./open-id-config');
 const bodyParser = require('body-parser');
